# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  service.UserRegisterService:
    properties:
      nickname:
        type: string
      password:
        type: string
      password_confirm:
        type: string
      user_name:
        type: string
    required:
    - nickname
    - password
    - password_confirm
    - user_name
    type: object
  service.UserUpdateService:
    properties:
      nickname:
        type: string
      password:
        type: string
    type: object
host: petstore.swagger.io:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /api/v1/user:
    delete:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: DeleteUser
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: GetUser
    put:
      parameters:
      - description: user info
        in: body
        name: user_info
        required: true
        schema:
          $ref: '#/definitions/service.UserUpdateService'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: EditUser
  /api/v1/user/register:
    post:
      parameters:
      - description: user info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/service.UserRegisterService'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.Response'
      summary: UserRegister
swagger: "2.0"